pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiida-trainspot"
dynamic = ["version"]
description = "AiiDA workchain to generate data for training and testing of neural networks."
authors = [{name = "Nataliia Manko"}, {name = "Davide Bidoggia"}, {name = "Antimo Marrazzo"}, {name = "Maria Peressi"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Development Status :: 4 - Beta",
    "Framework :: AiiDA"
]
keywords = ["aiida", "workchain"]
requires-python = ">=3.7"
dependencies = [
    "aiida-core>=2.6.2,<3",
    "voluptuous",
    "ase==3.22.1",
    "numba",
    "aiida-quantumespresso>=4.15.0",
    "aiida-lammps @ git+https://github.com/aiida-trieste-developers/aiida-lammps@ba554141e2ccb38de2b29a8f71427d115b74785a",
    "h5py"
]

[project.urls]
Source = "https://github.com/cuillier/aiida-trains-pot"

[project.entry-points.'aiida.workflows']
"trains_pot.datasetaugmentation"   = "aiida_trains_pot.workflows.datasetaugmentation:DatasetAugmentationWorkChain"
"trains_pot.macetrain"        = "aiida_trains_pot.workflows.mace_train:MaceTrainWorkChain"
"trains_pot.metatrain"        = "aiida_trains_pot.workflows.meta_train:MetaTrainWorkChain"
"trains_pot.labelling"        = "aiida_trains_pot.workflows.abinitiolabelling:AbInitioLabellingWorkChain"
"trains_pot.training"         = "aiida_trains_pot.workflows.training:TrainingWorkChain"
"trains_pot.workflow"         = "aiida_trains_pot.workflows.aiida_trains_pot:TrainsPotWorkChain"
"trains_pot.constrained"      = "aiida_trains_pot.workflows.pwconstrained:PwConstrainedWorkChain"
"trains_pot.lammpsmd"         = "aiida_trains_pot.workflows.lammpsmd_exploration:LammpsMDWorkChain"


[project.entry-points.'aiida.calculations']
"trains_pot.macetrain"        = "aiida_trains_pot.calculations.mace_train:MaceTrainCalculation"
"trains_pot.metatrain"        = "aiida_trains_pot.calculations.meta_train:MetaTrainCalculation"
"trains_pot.evaluation"       = "aiida_trains_pot.calculations.evaluation:EvaluationCalculation"
#"trains_pot.lammpsmd"         = "aiida_trains_pot.lammps.lammps_plugin.calculations:LammpsBaseCalculation"

[project.entry-points.'aiida.parsers']
"trains_pot.macetrain"        = "aiida_trains_pot.parsers.mace_train:MaceBaseParser"
"trains_pot.metatrain"        = "aiida_trains_pot.parsers.meta_train:MetaBaseParser"
"trains_pot.evaluation"       = "aiida_trains_pot.parsers.evaluation:EvaluationParser"
#"trains_pot.lammpsmd"         = "aiida_trains_pot.lammps.lammps_plugin.parsers:LammpsBaseParser"

[project.entry-points."aiida.data"]
"pesdata"                 = "aiida_trains_pot.data.pesdata:PESData"



##[project.entry-points."aiida.cmdline.data"]
[project.scripts]
portable_codes_installation = "aiida_trains_pot.utils.install_portable_codes:main"

[project.optional-dependencies]
docs = [
    "aiida-core~=2.5",
    "docutils",
    "sphinx",
    "sphinx-copybutton",
    "sphinx-book-theme",
    "sphinx-click",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe"
]

[tool.flit.module]
name = "aiida_trains_pot"

[tool.flit.sdist]
exclude = [
    "docs/",
    "tests/",
]



[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    "ignore:Object of type .* not in session, .* operation along .* will not proceed:sqlalchemy.exc.SAWarning",
    "ignore:Creating AiiDA configuration folder.*:UserWarning",
    "ignore::DeprecationWarning:frozendict:",
    "ignore::DeprecationWarning:pkg_resources:",
    "ignore::DeprecationWarning:sqlalchemy_utils:",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

format.quote-style = "double"
format.indent-style = "space"
format.skip-magic-trailing-comma = false
format.line-ending = "lf"

fix = true
show-fixes = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "PL"]

ignore = [
    "PLR0915",  # Too many statements
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments
    "D205",    # 1 blank line required between summary line and description
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
force-sort-within-sections = false
combine-as-imports = true
lines-between-types = 1

[tool.tox]
legacy_tox_ini = """
[tox]
envlist =
    py{38,39,310,311,312,313}
    py38-pre-commit

[testenv]
description = Run the pytest tests
usedevelop=True
extras = tests
commands = pytest {posargs}

[testenv:py38-pre-commit]
description = Run Ruff and pre-commit checks
extras =
    tests
    pre-commit
commands =
    pre-commit run --all-files
"""
